fix warnings

+++ test/UsageDemo.cpp
@@ -90,7 +90,7 @@ BOOST_AUTO_TEST_CASE(maeBlock)
             // not have the "i_m_template_index" property.
             const auto template_indices = atom_data->getIntProperty("i_m_template_index");
             for (size_t i=0; i<size; ++i) {
-                if (template_indices->isDefined(i)) {
+                if (template_indices->isDefined(static_cast<int>(i))) {
                     st->demo_property[i] = template_indices->at(i);
                 }
             }
